新增最大值最小值接口
# message auto-generated for no-merge-commit merge:
merge double into br_feature_NewFrameworkImpl_******

profiling数据解析，新增最大值最小值接口

Created-by: insightDeveloper

Match-id-a8a9d1a8d276918f5bc781610889c1b451a9f0bc

+++ server/src/modules/timeline/core/database/TraceDatabase.cpp
@@ -1196,7 +1196,7 @@ bool TraceDatabase::GetCommunicationDetails(const std::vector<std::string>& opTr
         Protocol::CommunicationDetail communicationDetail{};
         communicationDetail.communicationKernel = sqlite3_column_string(stmt, col++);
         double startTime = sqlite3_column_double(stmt, col++) -
-                static_cast<double>(Timeline::TraceTime::Instance().GetStartTime()) / 1000;
+                Timeline::TraceTime::Instance().GetStartTime() / 1000;
         double duration = sqlite3_column_double(stmt, col++);
         communicationDetail.startTime = startTime;
         communicationDetail.totalDuration = duration;
